+++ src/scriptlets/trusted-click-element.ts
@@ -24,14 +24,16 @@ import {
  * ```
  * <!-- markdownlint-disable-next-line line-length -->
  * - `selectors` — required, string with query selectors delimited by comma. The scriptlet supports `>>>` combinator to select elements inside open shadow DOM. For usage, see example below.
- * - `extraMatch` — optional, extra condition to check on a page; allows to match `cookie` and `localStorage`;
+ * - `extraMatch` — optional, extra condition to check on a page;
+ *    allows to match `cookie`, `localStorage` and specified text;
  * can be set as `name:key[=value]` where `value` is optional.
  * If `cookie`/`localStorage` starts with `!` then the element will only be clicked
- * if specified cookie/localStorage item does not exist.
+ * if specified `cookie`/`localStorage` item does not exist.
  * Multiple conditions are allowed inside one `extraMatch` but they should be delimited by comma
- * and each of them should match the syntax. Possible `name`s:
+ * and each of them should match the syntax. Possible `names`:
  *     - `cookie` — test string or regex against cookies on a page
  *     - `localStorage` — check if localStorage item is present
+ *     - `containsText` — check if clicked element contains specified text
  * - `delay` — optional, time in ms to delay scriptlet execution, defaults to instant execution.
  *
  * <!-- markdownlint-disable line-length -->
@@ -82,6 +84,12 @@ import {
  *     example.com#%#//scriptlet('trusted-click-element', 'button[name="agree"], input[type="submit"][value="akkoord"]', 'cookie:cmpconsent, localStorage:promo', '250')
  *     ```
  *
+ * 1. Click element only if clicked element contains text `Accept cookie`
+ *
+ *     ```adblock
+ *     example.com#%#//scriptlet('trusted-click-element', 'button', 'containsText:Accept cookie')
+ *     ```
+ *
  * 1. Click element only if cookie with name `cmpconsent` does not exist
  *
  *     ```adblock
@@ -105,7 +113,12 @@ import {
  * @added v1.7.3.
  */
 /* eslint-enable max-len */
-export function trustedClickElement(source, selectors, extraMatch = '', delay = NaN) {
+export function trustedClickElement(
+    source: Source,
+    selectors: string,
+    extraMatch = '',
+    delay = NaN,
+) {
     if (!selectors) {
         return;
     }
@@ -115,16 +128,17 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
     const STATIC_CLICK_DELAY_MS = 150;
     const COOKIE_MATCH_MARKER = 'cookie:';
     const LOCAL_STORAGE_MATCH_MARKER = 'localStorage:';
+    const TEXT_MATCH_MARKER = 'containsText:';
     const SELECTORS_DELIMITER = ',';
     const COOKIE_STRING_DELIMITER = ';';
     // Regex to split match pairs by commas, avoiding the ones included in regexes
-    const EXTRA_MATCH_DELIMITER = /(,\s*){1}(?=!?cookie:|!?localStorage:)/;
+    const EXTRA_MATCH_DELIMITER = /(,\s*){1}(?=!?cookie:|!?localStorage:|containsText:)/;
 
-    const sleep = (delayMs) => new Promise((resolve) => setTimeout(resolve, delayMs));
+    const sleep = (delayMs: number) => new Promise((resolve) => setTimeout(resolve, delayMs));
 
     let parsedDelay;
     if (delay) {
-        parsedDelay = parseInt(delay, 10);
+        parsedDelay = parseInt(String(delay), 10);
         const isValidDelay = !Number.isNaN(parsedDelay) || parsedDelay < OBSERVER_TIMEOUT_MS;
         if (!isValidDelay) {
             // eslint-disable-next-line max-len
@@ -136,8 +150,9 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
 
     let canClick = !parsedDelay;
 
-    const cookieMatches = [];
-    const localStorageMatches = [];
+    const cookieMatches: string[] = [];
+    const localStorageMatches: string[] = [];
+    let textMatches = '';
     let isInvertedMatchCookie = false;
     let isInvertedMatchLocalStorage = false;
 
@@ -161,6 +176,11 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
                 const localStorageMatch = matchValue.replace(LOCAL_STORAGE_MATCH_MARKER, '');
                 localStorageMatches.push(localStorageMatch);
             }
+            if (matchStr.includes(TEXT_MATCH_MARKER)) {
+                const { matchValue } = parseMatchArg(matchStr);
+                const textMatch = matchValue.replace(TEXT_MATCH_MARKER, '');
+                textMatches = textMatch;
+            }
         });
     }
 
@@ -179,8 +199,8 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
             const valueMatch = parsedCookieMatches[key] ? toRegExp(parsedCookieMatches[key]) : null;
             const keyMatch = toRegExp(key);
 
-            return cookieKeys.some((key) => {
-                const keysMatched = keyMatch.test(key);
+            return cookieKeys.some((cookieKey) => {
+                const keysMatched = keyMatch.test(cookieKey);
                 if (!keysMatched) {
                     return false;
                 }
@@ -190,7 +210,13 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
                     return true;
                 }
 
-                return valueMatch.test(parsedCookies[key]);
+                const parsedCookieValue = parsedCookies[cookieKey];
+
+                if (!parsedCookieValue) {
+                    return false;
+                }
+
+                return valueMatch.test(parsedCookieValue);
             });
         });
 
@@ -213,6 +239,26 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
         }
     }
 
+    const textMatchRegexp = textMatches ? toRegExp(textMatches) : null;
+
+    /**
+     * Checks if an element contains the specified text.
+     *
+     * @param element - The element to check.
+     * @param matchRegexp - The text to match.
+     * @returns True if the element contains the specified text, otherwise false.
+     */
+    const doesElementContainText = (
+        element: Element,
+        matchRegexp: RegExp,
+    ): boolean => {
+        const { textContent } = element;
+        if (!textContent) {
+            return false;
+        }
+        return matchRegexp.test(textContent);
+    };
+
     /**
      * Create selectors array and swap selectors to null on finding it's element
      *
@@ -221,17 +267,17 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
      * - always know on what index corresponding element should be put
      * - prevent selectors from being queried multiple times
      */
-    let selectorsSequence = selectors
+    let selectorsSequence: Array<string | null> = selectors
         .split(SELECTORS_DELIMITER)
         .map((selector) => selector.trim());
 
-    const createElementObj = (element) => {
+    const createElementObj = (element: any): Object => {
         return {
             element: element || null,
             clicked: false,
         };
     };
-    const elementsSequence = Array(selectorsSequence.length).fill(createElementObj());
+    const elementsSequence = Array(selectorsSequence.length).fill(createElementObj(null));
 
     /**
      * Go through elementsSequence from left to right, clicking on found elements
@@ -253,6 +299,9 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
             }
             // Skip already clicked elements
             if (!elementObj.clicked) {
+                if (textMatchRegexp && !doesElementContainText(elementObj.element, textMatchRegexp)) {
+                    continue;
+                }
                 elementObj.element.click();
                 elementObj.clicked = true;
             }
@@ -266,7 +315,7 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
         }
     };
 
-    const handleElement = (element, i) => {
+    const handleElement = (element: Element, i: number) => {
         const elementObj = createElementObj(element);
         elementsSequence[i] = elementObj;
 
@@ -283,8 +332,8 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
      * when delay is getting off after the last mutation took place.
      *
      */
-    const findElements = (mutations, observer) => {
-        const fulfilledSelectors = [];
+    const findElements = (mutations: MutationRecord[], observer: MutationObserver) => {
+        const fulfilledSelectors: string[] = [];
         selectorsSequence.forEach((selector, i) => {
             if (!selector) {
                 return;
@@ -300,7 +349,9 @@ export function trustedClickElement(source, selectors, extraMatch = '', delay = 
 
         // selectorsSequence should be modified after the loop to not break loop indexation
         selectorsSequence = selectorsSequence.map((selector) => {
-            return fulfilledSelectors.includes(selector) ? null : selector;
+            return selector && fulfilledSelectors.includes(selector)
+                ? null
+                : selector;
         });
 
         // Disconnect observer after finding all elements