+++ tests/scriptlets/trusted-click-element.test.js
@@ -16,11 +16,12 @@ const createSelectorsString = (clickOrder) => {
 };
 
 // Create clickable element with it's count as id and assertion as onclick
-const createClickable = (elementNum) => {
+const createClickable = (elementNum, text = '') => {
     const clickableId = `${CLICKABLE_NAME}${elementNum}`;
     const checkbox = document.createElement('input');
     checkbox.type = 'checkbox';
     checkbox.id = clickableId;
+    checkbox.textContent = text;
     checkbox.onclick = (e) => {
         e.currentTarget.setAttribute('clicked', true);
         window.clickOrder.push(elementNum);
@@ -229,6 +230,78 @@ test('extraMatch - single cookie match, matched', (assert) => {
     clearCookie(cookieKey1);
 });
 
+test('extraMatch - text match, matched', (assert) => {
+    const textToMatch = 'Accept cookie';
+    const EXTRA_MATCH_STR = `containsText:${textToMatch}`;
+
+    const ELEM_COUNT = 1;
+    // Check elements for being clicked and hit func execution
+    const ASSERTIONS = ELEM_COUNT + 1;
+    assert.expect(ASSERTIONS);
+    const done = assert.async();
+
+    const selectorsString = `#${PANEL_ID} > #${CLICKABLE_NAME}${ELEM_COUNT}`;
+
+    runScriptlet(name, [selectorsString, EXTRA_MATCH_STR]);
+    const panel = createPanel();
+    const clickable = createClickable(1, textToMatch);
+    panel.appendChild(clickable);
+
+    setTimeout(() => {
+        assert.ok(clickable.getAttribute('clicked'), 'Element should be clicked');
+        assert.strictEqual(window.hit, 'FIRED', 'hit func executed');
+        done();
+    }, 150);
+});
+
+test('extraMatch - text match regexp, matched', (assert) => {
+    const textToMatch = 'Reject foo bar cookie';
+    const EXTRA_MATCH_STR = 'containsText:/Reject.*cookie/';
+
+    const ELEM_COUNT = 1;
+    // Check elements for being clicked and hit func execution
+    const ASSERTIONS = ELEM_COUNT + 1;
+    assert.expect(ASSERTIONS);
+    const done = assert.async();
+
+    const selectorsString = `#${PANEL_ID} > #${CLICKABLE_NAME}${ELEM_COUNT}`;
+
+    runScriptlet(name, [selectorsString, EXTRA_MATCH_STR]);
+    const panel = createPanel();
+    const clickable = createClickable(1, textToMatch);
+    panel.appendChild(clickable);
+
+    setTimeout(() => {
+        assert.ok(clickable.getAttribute('clicked'), 'Element should be clicked');
+        assert.strictEqual(window.hit, 'FIRED', 'hit func executed');
+        done();
+    }, 150);
+});
+
+test('extraMatch - text match, not matched', (assert) => {
+    const textToMatch = 'foo';
+    const EXTRA_MATCH_STR = `containsText:${textToMatch}`;
+
+    const ELEM_COUNT = 1;
+    // Check elements for being clicked and hit func execution
+    const ASSERTIONS = ELEM_COUNT + 1;
+    assert.expect(ASSERTIONS);
+    const done = assert.async();
+
+    const selectorsString = `#${PANEL_ID} > #${CLICKABLE_NAME}${ELEM_COUNT}`;
+
+    runScriptlet(name, [selectorsString, EXTRA_MATCH_STR]);
+    const panel = createPanel();
+    const clickable = createClickable(1, 'bar');
+    panel.appendChild(clickable);
+
+    setTimeout(() => {
+        assert.notOk(clickable.getAttribute('clicked'), 'Element should not be clicked');
+        assert.strictEqual(window.hit, undefined, 'hit should not fire');
+        done();
+    }, 150);
+});
+
 test('extraMatch - single cookie match, not matched', (assert) => {
     const cookieKey1 = 'first';
     const cookieKey2 = 'second';
@@ -421,6 +494,31 @@ test('extraMatch - single cookie revert, click', (assert) => {
     runScriptlet(name, [selectorsString, EXTRA_MATCH_STR]);
     const panel = createPanel();
     const clickable = createClickable(1);
+    panel.appendChild(clickable);
+
+    setTimeout(() => {
+        assert.ok(clickable.getAttribute('clicked'), 'Element should be clicked');
+        assert.strictEqual(window.hit, 'FIRED', 'hit func executed');
+        done();
+    }, 150);
+});
+
+test('extraMatch - cookie revert + text match, click', (assert) => {
+    const textToMatch = 'Continue';
+    const cookieKey = 'revertTextTest';
+    const EXTRA_MATCH_STR = `!cookie:${cookieKey}, containsText:${textToMatch}`;
+
+    const ELEM_COUNT = 1;
+    // Check elements for being clicked and hit func execution
+    const ASSERTIONS = ELEM_COUNT + 1;
+    assert.expect(ASSERTIONS);
+    const done = assert.async();
+
+    const selectorsString = `#${PANEL_ID} > #${CLICKABLE_NAME}${ELEM_COUNT}`;
+
+    runScriptlet(name, [selectorsString, EXTRA_MATCH_STR]);
+    const panel = createPanel();
+    const clickable = createClickable(1, textToMatch);
     panel.appendChild(clickable);
 
     setTimeout(() => {